1.完成首页和数据看板界面的前后端开发 2.修改了部分文件的代码 3.添加了新表用于存储仓库每天的总价值和总数量

+++ backend/src/main/java/com/yunqi/backend/model/dto/RecordDTO.java
@@ -63,7 +63,7 @@ public class RecordDTO extends BaseDTO {
     /**
      * 总金额
      */
-    private int totalAmount;
+    private BigDecimal totalAmount;
 
     /**
      * 货物名称
